requirements.txt
qdrant-client==1.10.1
ragas==0.1.9
pandas==2.2.0
llama_index==0.10.56
selenium==4.21.0
openai==1.30.1
datasets==2.14.7
llama-index-llms-azure-openai==0.1.9
llama-index-embeddings-azure-openai==0.1.9
llama-index-llms-cohere 
llama-index-embeddings-cohere 
flagembedding==1.2.10
python-dotenv


